+++ Dreddy/WebContent/WEB-INF/jsp/adminpoliticaleventimg.jsp
@@ -203,7 +203,7 @@ Smartphone Compatible web template, free webdesigns for Nokia, Samsung, LG, Sony
 					</c:url>
 					<c:if test="${poliimglist.galleryimg_type=='political Events'}">
 					<div class="col-sm-4 gallery-grids-left">
-					<a href="${deleteadminpoliticalimg}"><i class="fa fa-trash" aria-hidden="true"></i></a>
+					<a href="${deleteadminpoliticalimg}" onclick="return confirmActiondelete();"><i class="fa fa-trash" aria-hidden="true"></i></a>
 				
 						<div class="gallery-grid">
 							<a class="example-image-link" href="images/galleryimages/${poliimglist.gallery_img}" data-lightbox="example-set" data-title="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Phasellus vitae cursus ligula">
@@ -330,6 +330,17 @@ function Validation()
 
 
 </script>
+
+<script type="text/javascript">
+function confirmActiondelete() {
+	if (!confirm("Are You Sure You want to Delete?")) {
+		return false;
+	}
+}
+
+
+</script>
+
 
 </body>
 </html>
